Refactor `ds_list_contains_ext`
The script now iterates through the provided values to check if they exist within the given DS list, instead of iterating through the list itself (this suggestion was made by Renex). The script's functionality has also been expanded by adding a `match_all` argument, which allows the user to check if all, or any one of, the provided values exist within the list.